+++ extension/src/editor/containers/common/fieldComponentEditor/fields/general/EditorLayerDescriptionField.tsx
@@ -5,7 +5,7 @@ import {
   PropertyBox,
   PropertyTextareaField,
   PropertyWrapper,
-} from "../../../../../../prototypes/ui-components";
+} from "../../../../ui-components/property";
 
 // TODO: Add markdown support
 export const EditorLayerDescriptionField: React.FC<BasicFieldProps<"LAYER_DESCRIPTION_FIELD">> = ({
