+++ extension/src/editor/containers/common/fieldComponentEditor/fields/general/EditorOpacityField.tsx
@@ -4,7 +4,7 @@ import {
   PropertyBox,
   PropertyPlaceHolder,
   NO_SETTINGS_FOR_THIS_COMPONNET,
-} from "../../../../../../prototypes/ui-components";
+} from "../../../../ui-components";
 
 export const EditorOpacityField: React.FC<BasicFieldProps<"OPACITY_FIELD">> = () => {
   return (
