+++ extension/src/prototypes/ui-components/index.ts
@@ -73,4 +73,3 @@ export * from "./styles";
 export * from "./SwitchParameterItem";
 export * from "./theme";
 export * from "./ValueParameterItem";
-export * from "../../editor/containers/ui-components";
